+++ src/Movim/Console/DaemonCommand.php
@@ -3,12 +3,14 @@
  * SPDX-FileCopyrightText: 2010 Jaussoin Timothée
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
+
 namespace Movim\Console;
 
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Helper\ProgressBar;
 
 use Respect\Validation\Validator;
 
@@ -58,8 +60,8 @@ class DaemonCommand extends Command
 
         if (config('daemon.url') && Validator::url()->notEmpty()->validate(config('daemon.url'))) {
             $baseuri = rtrim(config('daemon.url'), '/') . '/';
-        } elseif (file_exists(CACHE_PATH.'baseuri')) {
-            $baseuri = file_get_contents(CACHE_PATH.'baseuri');
+        } elseif (file_exists(CACHE_PATH . 'baseuri')) {
+            $baseuri = file_get_contents(CACHE_PATH . 'baseuri');
         } else {
             $output->writeln('<comment>Please load the login page once before starting the daemon to cache the public URL</comment>');
             $output->writeln('<comment>or configure DAEMON_URL in .env</comment>');
@@ -70,7 +72,7 @@ class DaemonCommand extends Command
             $output->writeln('<comment>Please set at least one user as an admin once its account is logged in</comment>');
 
             $output->writeln('<info>To set an existing user admin</info>');
-            $output->writeln('<info>php daemon.php setAdmin {jid}</info>'."\n");
+            $output->writeln('<info>php daemon.php setAdmin {jid}</info>' . "\n");
         }
 
         $compileLanguages = new \React\ChildProcess\Process('exec php daemon.php compileLanguages');
@@ -82,10 +84,33 @@ class DaemonCommand extends Command
         $compileStickers->on('exit', fn ($out) => $output->writeln('<info>Stickers compiled</info>'));
 
         $output->writeln('<info>Movim daemon launched</info>');
-        $output->writeln('<info>Base URL: '.$baseuri.'</info>');
+        $output->writeln('<info>Base URL: ' . $baseuri . '</info>');
 
         if ($input->getOption('debug')) {
-            $output->writeln("\n".'<comment>Debug is enabled, check the logs in syslog or '.DOCUMENT_ROOT.'/log/</comment>');
+            $output->writeln("\n" . '<comment>Debug is enabled, check the logs in syslog or ' . DOCUMENT_ROOT . '/log/</comment>');
+        }
+
+        $output->writeln("\n--- ".colorize("Opcache", 'purple')." ---");
+        if (isOpcacheEnabled()) {
+            $output->writeln('<info>Opcache is enabled compiling files...</info>');
+
+            $filesCount = count(listOpcacheCompilableFiles());
+
+            ProgressBar::setFormatDefinition('custom', '%current%/%max% | %message%');
+            $progressBar = new ProgressBar($output, $filesCount);
+            $progressBar->setFormat('custom');
+
+            foreach (compileOpcache() as $file) {
+                $progressBar->setMessage('Compiling...');
+                $progressBar->advance();
+            }
+
+            $progressBar->setMessage('Files compiled');
+            $progressBar->advance();
+            $output->writeln("");
+        } else {
+            $output->writeln('<error>Opcache is disabled, it is strongly advised to enable it in PHP CLI php.ini</error>');
+            $output->writeln('Set opcache.enable=1 and opcache.enable_cli=1');
         }
 
         $core = new Core($loop, $baseuri);